Added XboxService to Services and IXboxService to contracts plus edits to XboxModels and XboxGame entity

+++ Game.Models/XboxUpdateModel.cs
@@ -1,9 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel.DataAnnotations;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace Game.Models.XboxModels
 {
@@ -11,7 +6,7 @@ namespace Game.Models.XboxModels
     {
         public int XboxId { get; set; }
         public string Name { get; set; }
-        public decimal Price { get; set; }
+        public double Price { get; set; }
         public double Rating { get; set; }
     }
 }
